% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/emrlapply.R
\name{assembleOutput}
\alias{assembleOutput}
\title{Internal function used for assembling the output produced by the EMR process}
\usage{
assembleOutput(myTempDirOut)
}
\arguments{
\item{myTempDirOut}{the temp directory where the output files from the EMR
process are kept. Must be local, not S3.}
}
\value{
Returns a list of output that matches the input list in length
}
\description{
Internal function used for assembling the output produced by the EMR process
}
\author{
James "JD" Long
}

