% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/awsFunctions.R
\name{checkLastStepStatus}
\alias{checkLastStepStatus}
\title{AWS Support Function: Checks the status of a given job on EMR}
\usage{
checkLastStepStatus(jobFlowId)
}
\arguments{
\item{jobFlowId}{the Job Flow Id of the job to check}
}
\value{
Status of the last step
}
\description{
Checks the status of a previously issued step.
}
\author{
James "JD" Long
}

