% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/awsFunctions.R
\name{checkStatus}
\alias{checkStatus}
\title{AWS Support Function: Checks the status of a given job on EMR}
\usage{
checkStatus(jobFlowId)
}
\arguments{
\item{jobFlowId}{the Job Flow Id of the job to check}
}
\value{
Job Status
}
\description{
Checks the status of a previously issued job.
}
\author{
James "JD" Long
}

