% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/csvTransform.R
\name{csvToList}
\alias{csvToList}
\title{The inverse of the listToCsv() function
Takes a csv of serialized objects created by the listToCSV function and
turns it into a proper R list.}
\usage{
csvToList(inFileName)
}
\arguments{
\item{inFileName}{String pointing to the full path of the input CSV.}
}
\value{
Returns a list object. Or an error. Hopefully a list.
}
\description{
The inverse of the listToCsv() function
Takes a csv of serialized objects created by the listToCSV function and
turns it into a proper R list.
}
\examples{
myList <- NULL
  set.seed(1)
  for (i in 1:10){
    a <- c(rnorm(999), NA)
    myList[[i]] <- a
  }

  require(caTools)
  listToCsv(myList, "tst.csv")
  all.equal(myList,  csvToList("tst.csv" ))
}
\author{
James "JD" Long
}
\seealso{
listToCsv()
}

