% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/awsFunctions.R
\name{deleteS3Bucket}
\alias{deleteS3Bucket}
\title{AWS Support Function: Delete an S3 Bucket}
\usage{
deleteS3Bucket(bucketName)
}
\arguments{
\item{bucketName}{the bucket to be deleted}
}
\description{
Does nothing if the bucketName does not exist. If bucket contains Keys,
all keys are deleted.
}
\author{
James "JD" Long
}

