% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/awsFunctions.R
\name{deleteS3Key}
\alias{deleteS3Key}
\title{AWS Support Function: Delete an S3 Key (a.k.a file)}
\usage{
deleteS3Key(bucketName, keyName)
}
\arguments{
\item{bucketName}{name of the bucket}

\item{keyName}{the key in the bucket}
}
\description{
Deteles a key in a given bucket on S3
}
\author{
James "JD" Long
}

