% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/awsFunctions.R
\name{downloadS3File}
\alias{downloadS3File}
\title{AWS Support Function: Downloads a key from an S3 Bucket into a local file.}
\usage{
downloadS3File(bucketName, keyName, localFile)
}
\arguments{
\item{bucketName}{destination bucket}

\item{keyName}{key to download. ".all" to pull all keys}

\item{localFile}{local file name or path if ".all" is called for keyName}
}
\description{
Pulls a key (file) from a bucket into a localFile. If the keyName = ".all" then
all files from the bucket are pulled and localFile should be a
directory name. Ignores "sub directories" in buckets.
}
\author{
James "JD" Long
}

