% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/awsFunctions.R
\name{emptyS3Bucket}
\alias{emptyS3Bucket}
\title{AWS Support Function: Empty an S3 bucket}
\usage{
emptyS3Bucket(bucketName)
}
\arguments{
\item{bucketName}{Name of the bucket to be emptied}
}
\description{
Deletes all keys in the designated bucket
}
\author{
James "JD" Long
}

