\name{emrlapply-package}
\alias{emrlapply-package}
\docType{package}
\title{
Parallel lapply function for the Amazon Elastic Map Reduce (EMR) engine
}
\description{
Parallel lapply function for the Amazon Elastic Map Reduce (EMR) engine. This is NOT a map reduce project but it uses Hadoop Streaming on Amazon's EMR in order to get simple parallel computation ala cart. This is NOT a big data package. If you are CPU bound this might be a useful package. 
}
\details{
\tabular{ll}{
Package: \tab emrlapply\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2010-07-27\cr
License: \tab What license is it under?\cr
LazyLoad: \tab yes\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Who wrote it

Maintainer: Who to complain to <yourfault@somewhere.net>
~~ The author and/or maintainer of the package ~~
}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ package }

\examples{

}
