% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/emrlapply.R
\name{emrlapply}
\alias{emrlapply}
\title{Parallel lapply() function using Amazon's EMR service.}
\usage{
emrlapply(clusterObject, X, FUN, taskTimeout = 10, ...)
}
\arguments{
\item{clusterObject}{cluster on which to run the process}

\item{X}{list to which the function will be applied}

\item{FUN}{function to apply}

\item{taskTimeout}{maximum time a single unit of work can run (in minutes)}

\item{\dots}{other params to pass to FUN}
}
\value{
Output as a list
}
\description{
Parallel lapply() function for applying a function to every item in a list
using Amazon's EMR service.
}

