% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/csvTransform.R
\name{listToCsv}
\alias{listToCsv}
\title{Converts an R list into a text CSV.
Serializes each element of an R list into ASCII characters then encodes then
for use as the input to a Hadoop Streaming job.}
\usage{
listToCsv(inList, outFileName)
}
\arguments{
\item{inList}{list to convert to CSV}

\item{outFileName}{file name of resulting CSV file}
}
\value{
creates a CSV to file but returns nothing.
}
\description{
Converts an R list into a text CSV.
Serializes each element of an R list into ASCII characters then encodes then
for use as the input to a Hadoop Streaming job.
}
\author{
James "JD" Long
}
\seealso{
csvToList
}

