% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/awsFunctions.R
\name{makeS3Bucket}
\alias{makeS3Bucket}
\title{AWS Support Function: Creates an S3 Bucket}
\usage{
makeS3Bucket(bucketName)
}
\arguments{
\item{bucketName}{string of the name of the bucket to be created}
}
\description{
Creates an S3 bucket. If the bucket already exists, no warning is returned.
}
\author{
James "JD" Long
}

