% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/awsFunctions.R
\name{setCredentials}
\alias{setCredentials}
\title{##' AWS Support Function: set up credentials}
\usage{
setCredentials(awsAccessKeyText, awsSecretKeyText,
  setEnvironmentVariables = TRUE)
}
\arguments{
\item{awsAccessKeyText}{your AWS Access Key as a string}

\item{awsSecretKeyText}{your AWS Secret Key as a string}

\item{setEnvironmentVariables}{T/F would you like environment variables to be set so
Segue will read the credentials on load}
}
\description{
sets up the credentials needed to access AWS and optionally sets environment
variables for auto loading of credentials in the future
}
\author{
James "JD" Long
}

