% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/awsFunctions.R
\name{stopCluster}
\alias{stopCluster}
\title{Stops a running cluster}
\usage{
stopCluster(clusterObject)
}
\arguments{
\item{clusterObject}{a cluster object to stop}
}
\value{
nothing
}
\description{
Stops a running cluster and deletes temporary directories from EC2
}
\author{
James "JD" Long
}

