% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/awsFunctions.R
\name{submitJob}
\alias{submitJob}
\title{Submits a job to a running cluster}
\usage{
submitJob(clusterObject, stopClusterOnComplete = FALSE, taskTimeout = 10)
}
\arguments{
\item{clusterObject}{a cluster object to submit to}

\item{stopClusterOnComplete}{set to true if you want the cluster to be shut down
after job completes}

\item{taskTimeout}{maximum time a single unit of work can run (in minutes)}
}
\value{
Execution status of this job
}
\description{
After a cluster has been started this function submits jobs to that cluster.
If a job is submitted with enableDebugging=TRUE, all jobs submitted to that
cluster will also have debugging enabled. To turn debugging off, the cluster
must be stopped and restarted.
}

