\name{terminateCluster}
\alias{terminateCluster}
\title{Terminates a running clsuter.}
\usage{terminateCluster(clusterObject, deleteTemp=TRUE)}
\description{Terminates a running clsuter.
Terminates a running clsuter.}
\arguments{\item{clusterObject}{cluster to terminate}
\item{deleteTemp}{T/F whether to delete temp files}}
