% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/awsFunctions.R
\name{uploadS3File}
\alias{uploadS3File}
\title{AWS Support Function: Uploads a local file to an S3 Bucket}
\usage{
uploadS3File(bucketName, localFile)
}
\arguments{
\item{bucketName}{destination bucket}

\item{localFile}{local file to be uploaded}
}
\description{
If buckName does not exist, it is created and a warning is issued.
}
\author{
James "JD" Long
}

