\name{deleteS3Bucket}
\alias{deleteS3Bucket}
\title{AWS Support Function: Delete an S3 Bucket...}
\usage{deleteS3Bucket(bucketName)}
\description{AWS Support Function: Delete an S3 Bucket}
\details{Returns a warning if bucketName does not exist.}
\author{James Long}
\arguments{\item{bucketName}{the bucket to be deleted}}
