\name{emrlapply}
\alias{emrlapply}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Parallel lapply() function using Amazon's EMR service. 
}
\description{
Parallel lapply() function for applying a function to every item in a list using Amazon's EMR service. 
}
\usage{
emrlapply(X, FUN, clusterObject, ...)
}
\arguments{
  \item{X}{
}
  \item{FUN}{
}
  \item{clusterObject}{
}
  \item{\dots}{
}
}
\details{
}
\value{
}
\references{
}
\author{
}
\note{
}

\seealso{
}
\examples{

}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
