\name{emrlapply}
\alias{emrlapply}
\title{Parallel lapply() function using Amazon's EMR service.}
\usage{emrlapply(X, FUN, clusterObject, ...)}
\description{Parallel lapply() function using Amazon's EMR service.}
\details{Parallel lapply() function for applying a function to every item in a list
using Amazon's EMR service.}
\value{Output as a list}
\arguments{\item{X}{}
\item{FUN}{}
\item{clusterObject}{}
\item{\dots}{}}
