\name{listToCsv}
\alias{listToCsv}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Converts an R list into a text CSV.
}
\description{
Serializes each element of an R list into ASCII characters then encodes then for use as the input to a Hadoop Streaming job. 
}
\usage{
listToCsv(inList, outFileName)
}
\arguments{
  \item{inList}{
}
  \item{outFileName}{
}
}
\details{
}
\value{
}
\references{
}
\author{
  James JD Long
}
\note{
}

\seealso{
}
\examples{


}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
