\name{makeS3Bucket}
\alias{makeS3Bucket}
\title{AWS Support Function: Creates an S3 Bucket...}
\usage{makeS3Bucket(bucketName)}
\description{AWS Support Function: Creates an S3 Bucket}
\details{Returns a warning if bucketName already exists.}
\author{James "JD" Long}
\arguments{\item{bucketName}{string of the name of the bucket to be created}}
