\name{startCluster}
\alias{startCluster}
\title{
Starts a cluster on Amazon's EMR service
}
\description{
After a cluster has been defined with createCluster() this function actually starts the machines running. 
}
\usage{
startCluster(clusterObject)
}
\arguments{
  \item{clusterObject}{
}
}
\details{
}
\value{
}
\references{
}
\author{
}
\note{
}

\seealso{
}
\examples{


}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
