\name{stopCluster}
\alias{stopCluster}
\title{Stops a running cluster...}
\usage{stopCluster(clusterObject)}
\description{Stops a running cluster}
\details{Stops a running cluster - known as clusterFuck() and terminateCluster() in previous versions}
\value{not really sure - Jack Shit, I think}
\author{James "JD" Long}
\arguments{\item{clusterObject}{a cluster object to stop}}
