\name{submitJob}
\alias{submitJob}
\title{
Submits a job to a running cluster
}
\description{
Submits a job to a running cluster
}
\usage{
submitJob(clusterObject)
}
\arguments{
  \item{clusterObject}{
}
}
\details{
}
\value{
}
\references{
}
\author{
}
\note{
}

\seealso{
}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
