\name{terminateCluster}
\alias{terminateCluster}
\title{
Terminates a running clsuter. 
}
\description{
Terminates a running clsuter. 
}
\usage{
terminateCluster(clusterObject, deleteTemp = TRUE)
}
\arguments{
  \item{clusterObject}{
}
  \item{deleteTemp}{
}
}
\details{
}
\value{
}
\references{
}
\author{
}
\note{
}

\seealso{
}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
